+++ backend/internal/domains/story/types/types.go
@@ -1,7 +1,7 @@
 package types
 
 type Banner struct {
-	ID          int64  `json:"id,omitempty" db:"banner_id"`
+	ID          int    `json:"id,omitempty" db:"banner_id"`
 	Name        string `json:"name" db:"name"`
 	Description string `json:"description" db:"description"`
 	CreatedAt   string `json:"created_at,omitempty" db:"created_at"`
@@ -9,8 +9,8 @@ type Banner struct {
 }
 
 type Story struct {
-	ID        int64    `json:"id,omitempty" db:"stories_id"`
+	ID        int      `json:"id,omitempty" db:"stories_id"`
 	CreatedAt string   `json:"created_at,omitempty" db:"created_at"`
 	Banners   []Banner `json:"banners,omitempty" db:"banners"`
-	Creator   int64    `json:"creator,omitempty" db:"creator"`
+	Creator   int      `json:"creator,omitempty" db:"creator"`
 }
