+++ backend/internal/domains/user/types/types.go
@@ -1,7 +1,7 @@
 package types
 
 type User struct {
-	ID       int64  `json:"id,omitempty" db:"user_id"`
+	ID       int    `json:"id,omitempty" db:"user_id"`
 	Username string `json:"username" db:"username"`
 	Email    string `json:"email" db:"email"`
 	Avatar   string `json:"avatar" db:"avatar"`
